requirements.txt
# File: requirements.txt
# 祖龙 (ZULONG) 系统 Python 依赖

# 核心框架
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.40.0  # 4bit 量化

# 视觉处理
Pillow>=9.5.0
opencv-python>=4.7.0

# 自然语言处理
nltk>=3.8.0
jieba>=0.42.0

# 向量数据库
qdrant-client>=1.3.0
numpy>=1.24.0
scipy>=1.10.0

# Web 服务
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6

# 异步支持
aiohttp>=3.8.0
asyncio>=3.4.3

# 工作流编排（LangGraph 已用于 core/graph.py 和 l1/graph.py）
langgraph>=0.2.0

# MCP 协议支持（技能包生态接入）
mcp>=1.0.0

# 监控指标
prometheus-client>=0.17.0

# 系统监控
psutil>=5.9.0
pynvml>=11.5.0  # NVIDIA GPU 监控

# 日志
python-json-logger>=2.0.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 工具库
tqdm>=4.65.0
requests>=2.31.0

# 测试（开发环境）
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# 代码质量（开发环境）
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# 文档（开发环境）
mkdocs>=1.5.0
mkdocs-material>=9.0.0
